Document-Generation/backend/requirements.txt
Django==4.2.9
django-appconf==1.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.12.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.12.1
celery~=5.2.7
numpy==1.26.3
# psycopg2==2.9.5
psycopg2-binary==2.9.5
Pillow==9.3.0
django-cors-headers==3.13.0

